move ContractCallRewardCardButton to platform folder
